0.13.81_beta refactor

+++ appium_extended_exceptions/appium_extended_exceptions.py
@@ -279,7 +279,7 @@ class TapError(AppiumExtendedError):
                  y: int = None,
                  image: Union[bytes, str] = None,
                  duration: Optional[int] = None,
-                 timeout: int = 5,
+                 timeout: float = 5,
                  original_exception: Optional[Exception] = None
                  ):
         super().__init__(message)
